requirements.txt
fastapi
uvicorn[standard]
pydantic
boto3
python-jose[cryptography]
python-multipart
